+++ rust/main/chains/hyperlane-sealevel/src/mailbox.rs
@@ -50,7 +50,7 @@ use crate::{
     },
     SealevelKeypair,
 };
-use crate::{tx_submitter::TransactionSubmitter, utils::force_non_signers};
+use crate::{tx_submitter::TransactionSubmitter, utils::sanitize_dynamic_accounts};
 use crate::{ConnectionConf, SealevelProvider, SealevelRpcClient};
 
 const SYSTEM_PROGRAM: &str = "11111111111111111111111111111111";
@@ -90,11 +90,12 @@ pub struct SealevelMailbox {
 impl SealevelMailbox {
     /// Create a new sealevel mailbox
     pub fn new(
+        provider: SealevelProvider,
+        tx_submitter: Box<dyn TransactionSubmitter>,
         conf: &ConnectionConf,
-        locator: ContractLocator,
+        locator: &ContractLocator,
         payer: Option<SealevelKeypair>,
     ) -> ChainResult<Self> {
-        let provider = SealevelProvider::new(locator.domain.clone(), conf);
         let program_id = Pubkey::from(<[u8; 32]>::from(locator.address));
         let domain = locator.domain.id();
         let inbox = Pubkey::find_program_address(mailbox_inbox_pda_seeds!(), &program_id);
@@ -111,9 +112,7 @@ impl SealevelMailbox {
             outbox,
             payer,
             priority_fee_oracle: conf.priority_fee_oracle.create_oracle(),
-            tx_submitter: conf
-                .transaction_submitter
-                .create_submitter(provider.rpc().url()),
+            tx_submitter,
             provider,
         })
     }
@@ -286,9 +285,8 @@ impl SealevelMailbox {
 
         let account_metas = self.get_account_metas(instruction).await?;
 
-        // Force all dynamically provided account metas to be non-signers to protect against
-        // potential theft from the payer.
-        Ok(force_non_signers(account_metas))
+        // Ensure dynamically provided account metas are safe to prevent theft from the payer.
+        sanitize_dynamic_accounts(account_metas, &self.get_payer()?.pubkey())
     }
 
     async fn get_process_instruction(
@@ -576,12 +574,15 @@ pub struct SealevelMailboxIndexer {
 impl SealevelMailboxIndexer {
     /// Create a new SealevelMailboxIndexer
     pub fn new(
+        provider: SealevelProvider,
+        tx_submitter: Box<dyn TransactionSubmitter>,
+        locator: &ContractLocator,
         conf: &ConnectionConf,
-        locator: ContractLocator,
         advanced_log_meta: bool,
     ) -> ChainResult<Self> {
+        let mailbox = SealevelMailbox::new(provider, tx_submitter, conf, locator, None)?;
+
         let program_id = Pubkey::from(<[u8; 32]>::from(locator.address));
-        let mailbox = SealevelMailbox::new(conf, locator, None)?;
 
         let dispatch_message_log_meta_composer = LogMetaComposer::new(
             mailbox.program_id,