+++ typescript/infra/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @hyperlane-xyz/infra
 
+## 8.8.0
+
+### Patch Changes
+
+- Updated dependencies [719d022ec]
+- Updated dependencies [c61546cb7]
+- Updated dependencies [d82d24cc7]
+- Updated dependencies [b054b0424]
+  - @hyperlane-xyz/sdk@8.8.0
+  - @hyperlane-xyz/cli@8.8.0
+  - @hyperlane-xyz/helloworld@8.8.0
+  - @hyperlane-xyz/utils@8.8.0
+
+## 8.7.0
+
+### Minor Changes
+
+- db832b803: Added support for multiple registries in CLI with prioritization.
+
+### Patch Changes
+
+- Updated dependencies [bd0b8861f]
+- Updated dependencies [55db270e3]
+- Updated dependencies [b92eb1b57]
+- Updated dependencies [db832b803]
+- Updated dependencies [ede0cbc15]
+- Updated dependencies [12e3c4da0]
+- Updated dependencies [d6724c4c3]
+- Updated dependencies [7dd1f64a6]
+- Updated dependencies [d93a38cab]
+  - @hyperlane-xyz/sdk@8.7.0
+  - @hyperlane-xyz/cli@8.7.0
+  - @hyperlane-xyz/helloworld@8.7.0
+  - @hyperlane-xyz/utils@8.7.0
+
 ## 8.6.1
 
 ### Patch Changes